+++ db/migrate/20111103005304_add_hyp_days_to_calf.rb
@@ -1,6 +1,6 @@
 class AddHypDaysToCalf < ActiveRecord::Migration
   def self.up
-    add_column :calves, :hypDays, :double
+    add_column :calves, :hypDays, :integer
   end
 
   def self.down
